13 - Capstone RAG/requirements.txt
langchain==0.3.0
faiss-cpu==1.10.0
openai==1.70.0
pypdf==3.17.0
tiktoken==0.9.0
python-dotenv==1.1.0
langchain-openai==0.2.0
langchain-community==0.3.0
langchain-core==0.3.0
pydantic==2.7.4
jira==3.8.0
gradio==4.44.1